+++ convert.go
@@ -2,11 +2,22 @@ package main
 
 import (
 	"io"
+	"os"
+
+	log "github.com/sirupsen/logrus"
 
 	report "gitlab.com/gitlab-org/security-products/analyzers/report/v2"
 	"gitlab.com/gitlab-org/security-products/analyzers/semgrep/sarif"
 )
 
 func convert(reader io.Reader, prependPath string) (*report.Report, error) {
-	return sarif.TransformToGLSASTReport(reader, prependPath)
+	// HACK: extract root path from environment variables
+	// TODO: https://gitlab.com/gitlab-org/gitlab/-/issues/320975
+	root := os.Getenv("ANALYZER_TARGET_DIR")
+	if root == "" {
+		root = os.Getenv("CI_PROJECT_DIR")
+	}
+
+	log.Debugf("Converting report with the root path: %s", root)
+	return sarif.TransformToGLSASTReport(reader, root)
 }