+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 Semgrep analyzer changelog
 
+## v0.4.0
+- Fix paths in report to be relative to project root (!6)
+- Update Dockerfile to give us control over base image (!6)
+- Update sarif package to support multiple runs and locations (!6)
+
 ## v0.3.0
 - OWASP metadata added to bandit rule-set (!9)